+++ apps/sim/lib/table/__tests__/update-row.test.ts
@@ -3,6 +3,7 @@
  */
 import { dbChainMock, dbChainMockFns, resetDbChainMock } from '@sim/testing'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
+import { deleteColumn, renameColumn } from '@/lib/table/columns/service'
 import {
   batchInsertRows,
   insertRow,
@@ -10,7 +11,6 @@ import {
   updateRow,
   upsertRow,
 } from '@/lib/table/rows/service'
-import { deleteColumn, renameColumn } from '@/lib/table/service'
 import type { TableDefinition } from '@/lib/table/types'
 import { getUniqueColumns } from '@/lib/table/validation'
