+++ apps/sim/lib/table/index.ts
@@ -7,6 +7,7 @@
 
 export * from './billing'
 export * from './column-keys'
+export * from './columns/service'
 export * from './constants'
 export * from './import'
 export * from './llm'
